restoring concept of entries internally
this change is largely to help make working with trees in parallel easier, as it removes
some of the more magical/destructive parts of the code.

entries are those files that come from builders, and as treated as special-cases so that
they won't be automatically pruned from the tree, such as when you are only building files
on-demand in a server.