fix: properly handle files without an extension (#23)
* test: add coverage for cases with dot in dirname and files without an extension

* fix: use basename for determining file type

The previous logic used `file.path` when checking type, which allowed a file that had no extension, but was in a directory that included a `.` to report an unexpected type.

For example: `dominicbarnes.github.com/CNAME` showed the type `com/cname`, which makes no sense

The change here switches to using `file.basename`, which by itself means the example above returns `""`. The next step is to fall back to a lowercased `file.basename` to get the expected result of `cname`.